refactored schema

+++ db/migrate/20190918201601_create_notes.rb
@@ -4,8 +4,8 @@ class CreateNotes < ActiveRecord::Migration[5.2]
       t.belongs_to :kit, foreign_key: true
       t.belongs_to :user, foreign_key: true
       t.text :note
-      t.integer :note_pad_start
-      t.integer :note_pad_end
+      t.float :note_pad_start
+      t.float :note_pad_end
 
       t.timestamps
     end
